Renombre de imágenes para evitar errores ortográficos y arreglo de algunas cosas respecto a la manera de escritura

+++ src/views/event/EventPage.vue
@@ -70,6 +70,7 @@ const modalDeleteInfo = ref({
     closeText: 'Cancelar',
     acceptText: 'Eliminar',
     onAccept: async () => {
+        if (!isAdmin.value) return;
         await deleteEvent(event.value._id);
         router.push('/events');
     }
